feat: add entity design
