+++ src/main/java/com/socialcommerce/SocialcommerceDomainApplication.java
@@ -1,4 +1,4 @@
-package com.socialcommerce.domain;
+package com.socialcommerce;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
