+++ src/main/java/com/socialcommerce/common/enums/PurchaseStatus.java
@@ -1,4 +1,4 @@
-package com.socialcommerce.domain.common.enums;
+package com.socialcommerce.common.enums;
 
 public enum PurchaseStatus {
   PENDING,
