+++ src/main/java/com/socialcommerce/common/enums/RefundType.java
@@ -1,0 +1,6 @@
+package com.socialcommerce.common.enums;
+
+public enum RefundType {
+    ALL, PARTIAL
+}
+
