+++ src/main/java/com/socialcommerce/domain/category/entity/Category.java
@@ -1,0 +1,65 @@
+package com.socialcommerce.domain.category.entity;
+
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.DynamicInsert;
+import org.hibernate.annotations.DynamicUpdate;
+import org.hibernate.annotations.UpdateTimestamp;
+
+@Entity
+@Table
+@Getter
+@DynamicInsert
+@DynamicUpdate
+@NoArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class Category {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  Long id;
+
+  @Column(nullable = false)
+  String name;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "parent_id")
+  @JsonBackReference  // 만약 Category 를 클라이언트에서 API 로 조회를 해서 Response 로 보내줄때 Json 으로 바뀌면서 뽑을때 순환참조를 하게 됨을 막아준다.
+  Category parent;
+
+  @OneToMany(mappedBy = "parent", fetch = FetchType.LAZY)
+  @JsonManagedReference
+  List<Category> children = new ArrayList<>();
+
+  @CreationTimestamp
+  @Column(nullable = false, updatable = false)
+  LocalDateTime createdAt;
+  @UpdateTimestamp
+  @Column
+  LocalDateTime updatedAt;
+
+  @Builder
+  public Category(String name, Category parent) {
+    this.name = name;
+    this.parent = parent;
+  }
+}
