--- src/main/java/com/socialcommerce/domain/common/enums/RefundType.java
@@ -1,6 +1,0 @@
-package com.socialcommerce.domain.common.enums;
-
-public enum RefundType {
-    ALL, PARTIAL
-}
-
