+++ src/main/java/com/socialcommerce/domain/domain/product/entity/Product.java
@@ -1,75 +1,5 @@
 package com.socialcommerce.domain.domain.product.entity;
 
+public class Product {
 
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import com.socialcommerce.domain.domain.cart.entity.Cart;
-import com.socialcommerce.domain.domain.category.entity.Category;
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import java.util.List;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.FieldDefaults;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.DynamicInsert;
-import org.hibernate.annotations.DynamicUpdate;
-import org.hibernate.annotations.UpdateTimestamp;
-
-@Entity
-@Table
-@Getter
-@DynamicInsert
-@DynamicUpdate
-@NoArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class Product {  // 상품
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "category_id")
-  private Category category;
-
-  @Column(nullable = false)
-  private String name;
-
-  @Column(columnDefinition = "TEXT")
-  private String description;
-
-  @Column(nullable = false)
-  private BigDecimal price;
-
-  @Column(nullable = false)
-  private Integer stock;
-
-  @CreationTimestamp
-  @Column(nullable = false, updatable = false)
-  LocalDateTime createdAt;
-
-  @UpdateTimestamp
-  @Column
-  LocalDateTime updatedAt;
-
-  @Builder
-  public Product(String name, String description, BigDecimal price, Integer stock) {
-    this.name = name;
-    this.description = description;
-    this.price = price;
-    this.stock = stock;
-  }
 }
