+++ src/main/java/com/socialcommerce/domain/refund/entity/Refund.java
@@ -1,0 +1,73 @@
+package com.socialcommerce.domain.refund.entity;
+
+import com.socialcommerce.common.enums.RefundStatus;
+import com.socialcommerce.common.enums.RefundType;
+import com.socialcommerce.domain.purchase.entity.Purchase;
+import com.socialcommerce.domain.purchase.entity.PurchaseProduct;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.FieldDefaults;
+import org.hibernate.annotations.DynamicInsert;
+import org.hibernate.annotations.DynamicUpdate;
+@Getter
+@Entity
+@Table(name = "refunds")
+@DynamicInsert
+@DynamicUpdate
+@NoArgsConstructor  // 빈생성자를 만들어준다. 항상 있는데 귀찮기 때문에
+@FieldDefaults(level = AccessLevel.PRIVATE) // 모든 접근 제한이 private 으로 바뀐다.
+public class Refund {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    Long id;
+
+    // 전체환불
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "purchase_id", nullable = true)
+    private Purchase purchase;
+
+    // 부분환불
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "purchase_product_id", nullable = false, unique = true)
+    private PurchaseProduct purchaseProduct;
+
+    @Column(nullable = false, length = 255)
+    String reason;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false, length = 20)
+    RefundStatus status;
+
+    // 전체환불 or 부분환불 Enum
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false, length = 10)
+    private RefundType type; // ALL or PARTIAL
+
+    @Column(name = "created_at", updatable = false)
+    LocalDateTime createdAt;
+
+    @Column(name = "updated_at")
+    LocalDateTime updatedAt;
+
+    @Builder
+    public Refund(PurchaseProduct purchaseProduct, String reason, RefundStatus status) {
+        this.purchaseProduct = purchaseProduct;
+        this.reason = reason;
+        this.status = status;
+    }
+}
+
