+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 spring:
   # 데이터베이스 연결 설정
   datasource:
-    url: jdbc:mysql://localhost:3306/spring_yap_db?useSSL=false&allowPublicKeyRetrieval=true
+    url: jdbc:mysql://localhost:3307/spring_yap_db?useSSL=false&allowPublicKeyRetrieval=true
     username: root
     password: root
     driver-class-name: com.mysql.cj.jdbc.Driver
