+++ src/test/java/com/socialcommerce/SocialcommerceDomainApplicationTests.java
@@ -1,4 +1,4 @@
-package com.socialcommerce.domain;
+package com.socialcommerce;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
